requirements.txt
# Core LangChain packages
langchain==0.3.3
langchain-community==0.3.2
langchain-core==0.3.10
langchain-upstage==0.2.0

# Vector databases
chromadb==0.5.3
faiss-cpu==1.9.0.post1

# Database
sqlalchemy==2.0.30
pymysql==1.1.1
psycopg2-binary==2.9.9

# Web scraping
beautifulsoup4==4.12.3
requests==2.32.3
lxml==5.2.2
html5lib==1.1

# Environment and configuration
python-dotenv==1.0.1
pyyaml==6.0.1

# UI frameworks
streamlit==1.37.1
fastapi==0.111.0
uvicorn[standard]==0.30.1

# Utilities
tqdm==4.66.4
pandas==2.2.2
numpy==1.26.4
aiofiles==23.2.1

# Development tools
pytest==8.2.2
pytest-cov==5.0.0
pytest-asyncio==0.23.7
black==24.4.2
flake8==7.0.0
ipython==8.26.0
jupyter==1.0.0

# Monitoring and logging
loguru==0.7.2

# Optional: Performance monitoring
# prometheus-client==0.20.0
# grafana-client==3.5.0